delete user service

+++ internal/user/repository.go
@@ -1,6 +1,10 @@
 package user
 
-import "github.com/ShopOnGO/ShopOnGO/prod/pkg/db"
+import (
+	"errors"
+
+	"github.com/ShopOnGO/ShopOnGO/prod/pkg/db"
+)
 
 type UserRepository struct {
 	Database *db.Db
@@ -14,6 +18,9 @@ func NewUserRepository(database *db.Db) *UserRepository {
 
 func (repo *UserRepository) Create(user *User) (*User, error) {
 	result := repo.Database.DB.Create(user)
+	if user.Email == "" {
+		return nil, errors.New("email is required")
+	}
 	if result.Error != nil {
 		return nil, result.Error
 	}
@@ -23,6 +30,9 @@ func (repo *UserRepository) Create(user *User) (*User, error) {
 }
 
 func (repo *UserRepository) FindByEmail(email string) (*User, error) {
+	if email == "" {
+		return nil, errors.New("email cannot be empty")
+	}
 	var user User
 	result := repo.Database.DB.First(&user, "email = ?", email) // SQL QUERY BY CONDS
 	if result.Error != nil {
@@ -32,6 +42,9 @@ func (repo *UserRepository) FindByEmail(email string) (*User, error) {
 }
 
 func (repo *UserRepository) Update(user *User) (*User, error) {
+	if user.ID == 0 {
+		return nil, errors.New("user ID is required for update")
+	}
 	result := repo.Database.DB.Model(&User{}).Where("id = ?", user.ID).Updates(user)
 	if result.Error != nil {
 		return nil, result.Error
@@ -40,6 +53,9 @@ func (repo *UserRepository) Update(user *User) (*User, error) {
 }
 
 func (repo *UserRepository) Delete(id uint) error {
+	if id == 0 {
+		return errors.New("user ID is required for deletion")
+	}
 	result := repo.Database.DB.Delete(&User{}, id)
 	if result.Error != nil {
 		return result.Error
@@ -48,12 +64,18 @@ func (repo *UserRepository) Delete(id uint) error {
 }
 
 func (repo *UserRepository) UpdateUserPassword(id uint, newPassword string) error {
+	if id == 0 {
+		return errors.New("user ID is required for UpdateUserPassword")
+	}
 	result := repo.Database.DB.Model(&User{}).Where("id = ?", id).Update("password", newPassword)
 	return result.Error
 }
 
 func (repo *UserRepository) GetUserRoleByEmail(email string) (string, error) {
-    var role string
+	if email == "" {
+		return "", errors.New("email cannot be empty")
+	}
+	var role string
     result := repo.Database.DB.Model(&User{}).Select("role").Where("email = ?", email).First(&role)
     if result.Error != nil {
         return "", result.Error
@@ -62,6 +84,9 @@ func (repo *UserRepository) GetUserRoleByEmail(email string) (string, error) {
 }
 
 func (repo *UserRepository) UpdateRole(user *User, newRole string) error {
-    result := repo.Database.DB.Model(&User{}).Where("id = ?", user.ID).Update("role", newRole)
+    if user.ID == 0 {
+		return errors.New("user ID is required for UpdateUserPassword")
+	}
+	result := repo.Database.DB.Model(&User{}).Where("id = ?", user.ID).Update("role", newRole)
     return result.Error
 }
